DON-887: Take select fields as array not object to be able to display in desired order
Iterating over an object doesn't usefully have a well defined order. As
the order we present options matters, we need to have the options passed
in as an array, not just an object.

This is a BC break - I will pull it into donate imediatly and update the
calling code there. We only currently use biggive-form-field-select in
one place in donate, and it is not used in wordpress.

See https://stackoverflow.com/questions/5525795/does-javascript-guarantee-object-property-order