WPD-162: Update readme to check how this affects wordpress
As few changes as possible in a commit. I want to see whether the hash
mentioned in dist/esm/loader.js changes from a19ad098 and how pulling this into
wordpress affects things.

Apparently it does not change - but then as no actual js files are
different that may not matter.