WPD-153: Make Form Field Select element 'Form Associated'
2nd attempt at making this element usable in an HTML form. Passing name
through doesn't work because the stuff in the shadow dom can't be
submitted as part of the form.

Based on docs at https://stenciljs.com/docs/form-associated

Browser support is not perfect but I think good enough if we no longer
need to support IE 11: https://caniuse.com/?search=attachInternals

And sadly using attach internals seems to mean we can't unit test this -
we were getting the error below. But maybe we don't need tests for now,
at least while we're not developing any complex logic in this component.
We still have the e2e test.

	FAIL src/components/biggive-form-field-select/test/biggive-form-field-select.spec.tsx
	  ● Console

	    console.error
	      NOTE: Property setFormValue was accessed on ElementInternals, but this property is not implemented.
	      Testing components with ElementInternals is fully supported in e2e tests.

	      53 |
	      54 |   componentWillLoad() {
	    > 55 |     this.internals.setFormValue(this.selectedValue);
		 |                    ^
	      56 |   }
	      57 |
	      58 |   render() {