DON-970: Display "Relevance" as an option in campagin card filter grid when a search term is present
This seems to work in testing with the index.html in the components
repo, but I'd be more confident after testing it with the
donate-frontend. As this is quite a complex component that is IMHO
unlikely to ever have an application outside of donate frontend I think
we'd be better off with it defined inside that application instead of
here. Not necassarily worth moving unless we're going to do more work on
it though.