DON-961: Fix hero image layout issue in server side render
Currently when we do a server side render something (not sure what) that
I think is part of our Stencil library inserts whitespace and comments
that show the HYDRATE_CHILD_ID / c-id of various elements, e.g.:

                    <div c-id="44702.7.3.2" class="teaser">
                      <!--t.44702.8.4.0-->
                      Hi. We’re Big Give. Our match funding campaigns offer a unique opportunity for you to double the difference you can make in the world for the causes close to your heart.
                    </div>

This combined with our use of `whitespace: pre-wrap` removed here meant
the layout of our main title and subtitle on the homepage was messed up
in the SSR version of the page, and visible moved between rendering that
and the client side rendered version.

So removing `pre-wrap` here and replacing with <br/> tags.