DON-998: Make "more info" button optional on campaign cards
As we currently don't have a campaign page for individual regular giving
campaigns, there isn't always a place to link to. When we're not on the
charity's page we link there, but when this is displayed on the actual
charity page there's no obvious place that this should link to.