COM-104: Fix philco quote layout
Previous version of Philco Quote SVG was wider than content which meant
it was not able to center properly in its container. I edited it in
inkscape to resize the svg to fit content.

This extra SVG width was also making the quote underneath too wide
within a grid and either overlap with its neighbour or spread past the
edge of a screen on a narrow screen viewport