WPD-197: Make biggive-formatted-text behaviour the same when given no defaultTextColour as when given blank defaultTextColour
WP currently passes an empty string as defaultTextColour in many cases. To avoid
relying on that accidental behaviour change makes the component behave
the same way if the defaultTextColour is not passed at all, instead of
defaulting the text to blue (aka primary) in that case as it does now.

As far as I can tell WP never currently passes an undefined value here,
and in frontend it is only used to wrap a link which has its own colour
defined, so this is should not be a BC break in any way that matters.

Only usages in FE:
https://github.com/thebiggive/donate-frontend/blob/1756d9c02b0e9b50af5a69ee3e90993f506e6acb/src/app/campaign-details/campaign-details.component.html#L15

The cases where no `default-text-colour` are either just wrapping links,
which have their own colour set, or in the case of `Updates` are
probably just blue by mistake and we'd rather have the updates in black
anyway for consistency.

And the only usage in WP currently cooerces the value to string to
passes empty string and never undefined. With empty string the class
currently becomes "text-colour-" which is meaningless so acts the same
as no class at all.