DON-1120: Remove height limit for banner text, take three separate colours
Height will now be whatever's needed to fit the content. We trust
content authors not to put too much content in.

Text may be any three colours as chosen in WP. For now in frontend the
three pieces of text will all be the same colour.

BC break but should be detectable by SA in frontend.